pyproject.toml
[tool.poetry]
name = "caracal"
version = "1.1.5"
description = "Containerized Automated Radio Astronomy Calibration"
authors = ["The CaraCAL Team <caracal-info@googlegroups.com>"]
readme = "README.rst"
homepage = "https://github.com/caracal-pipeline/caracal"
repository = "https://github.com/caracal-pipeline/caracal"
keywords = ["Astronomy", "Visualisation", "Pipelines"]
classifiers = [
        "Development Status :: 5 - Production/Stable",
        "Intended Audience :: Science/Research",
        "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
        "Operating System :: POSIX :: Linux",
        "Programming Language :: Python",
        "Topic :: Scientific/Engineering :: Astronomy"
]
license = "GNU GPL v2"
packages = [
    {include = "caracal"},
]

[tool.poetry.scripts]
caracal = "caracal.main:driver"

[tool.poetry.dependencies]
python = ">=3.9,<=3.13"
stimela = {git="https://github.com/ratt-ru/Stimela-classic@master"}
psutil = "*"
pykwalify = "*"
progressbar2 = "*"
"ruamel.yaml" = "*"
future_fstrings = "*"
astropy = [
    { version = "5.3.3", python = ">=3.9,<3.11"},
    {version= "^7.0", python=">=3.11,<=3.13"}
]
uritools = "*"
scipy = "*"
regions = "*"
astroquery = "*"
numpy = "*"

# These are not essentail for the pipeline to run, so better to not pin the versions.
nbconvert = "*"
radiopadre-client = "*"
python-casacore = "*"
matplotlib = "*"
jinja2 = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
scipy = ["scipy"]
astropy = ["astropy"]
regions = ["regions"]
astroquery = ["astroquery"]
all = ["scipy", "astropy", "regions", "astroquery"]


[tool.poetry.group.tests]
optional = true
[tool.poetry.group.tests.dependencies]
pytest = "*"
flake8 = "*"

[tool.poetry.group.docs]
optional = true
[tool.poetry.group.docs.dependencies]
Sphinx = "^4.0.1"
sphinx-copybutton = "^0.5.0"
furo = "^2022.9.15"
